requirements.txt
fastapi
uvicorn
openai
pydantic
python-dotenv
pypdf
python-multipart
langchain-openai
langchain-core
langchain_google_genai
mcp
contextlib
asyncio